requirements.txt
absl-py==2.1.0
accelerate==1.2.0
aif360==0.6.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
BlackBoxAuditing==0.1.54
blinker==1.9.0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
datasets==3.1.0
dill==0.3.8
fairlearn==0.12.0
Flask==3.1.0
flatbuffers==25.1.24
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.70.0
h5py==3.12.1
huggingface-hub==0.26.5
imageio==2.37.0
imbalanced-learn==0.13.0
imblearn==0.0
inFairness==0.2.3
itsdangerous==2.2.0
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
lime==0.2.0.1
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
matplotlib==3.10.0
mdurl==0.1.2
mkl-service==2.4.0
ml-dtypes==0.4.1
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
numba==0.61.0
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
POT==0.9.5
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.0
pyarrow==18.1.0
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.11.6
rich==13.9.4
safetensors==0.4.5
scikit-image==0.25.2
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
setuptools==75.1.0
shap==0.47.0
six==1.17.0
sklearn-compat==0.1.3
skorch==1.1.0
slicer==0.0.8
soupsieve==2.6
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.2.18
tokenizers==0.21.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.47.0
tzdata==2024.2
ucimlrepo==0.0.7
Werkzeug==3.1.3
wheel==0.44.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
